rename `ratio` to `scale` for splits
Naming this property `ratio` does not make a whole lot of sense. Hence we rename
it to `scale` and also make it reflect the scaling of the main side of the split.
This effectively means one should subtract the current value from 1.0 to convert
it to the new scaling value (e.g. 0.25 becomes 0.75).