refactor migration code
Migrating views from one output to another during output destroy turns out to be
a lot trickier. The commit logic needs to be adjusted as well to prevent views
from jumping around on show. Also view geometry needs to be reset and therefore
tiles need to be detached from layouts, which leads to layout deallocation when
there are no more views left inside the layout.