streamline cycling
Since layout mode the default bindings for layout cycling have been bound to
`L+n` and `L+p` by default. Cycling through views can be done with `LS+n` and
`LS+p.` This leads to muscle memory issues when going to the "previous" view
because `hikari` interprets the next view as the one going down the stack and
the previous one as going up the stack. This is counterintuitive because going
back to the previous view requires the user to issue `LS+n`. Therefore we just
switch the meaning for next and previous around for everything by layout
cycling.

Also cycling through sheet views only can be very disorienting with borrowed
views on the workspace. Therefore we replace sheet cycling with cycling though
all the views on the workspace.