handle reset during pending tiling
Resetting geometry should always work even when the view is marked as dirty.
This means that there can be a pending tile structure which would not be
properly freed in this case. We now force cancellation of the tiling operation
and free the pending tile in the process. This is the same operation that
happens when we destroy a view with a pending tile.