deprecate release bindings for action begin/end
Binding actions to the release event for a certain key is not enough for
functionality like "push to talk". The release action might get lost when a user
releases modifiers first.

This patch introduces *begin* and *end* actions. The *begin* action is followed
by the *end* action once we release any key or push more keys.