remove sheet groups
Sheets no longer have a special group. Each view needs to be part of a group
though, therefore we take the app id or class of a view when no explicit group
is assigned. This conveniently allows the user to figure out app id or class of
a view without the need for additional tooling.