handle layer destroy before map
As documented in issue #1 destroying a layer before mapping it leads to the
freed memory still being linked to the `layer_surfaces` list. This patch adds
the invariant the we always need to be able to remove this link. This is ensured
by doing a `wl_list_init` after every `wl_list_remove`.