Makefile changes to support more general builds
Pull ${CFLAGS_EXTRA} and ${LDFLAGS_EXTRA} from the environment, which allows the
user to specify additional CFLAGS and LDFLAGS that will be added to the
respective CFLAGS and LDFLAGS variables built up by the Makefile. (Setting
CFLAGS or LDFLAGS directly on the command line, at least with NetBSD make, would
cause make to ignore all of the appends in the Makefile.)

Add ${CFLAGS_EXTRA} and ${LDFLAGS_EXTRA} to the hikari-unlocker build command,
so those flags apply when building that target (this is important in Void
cross-compilation, because PAM libs and headers will not generally be in
/usr/lib or /usr/include and require specification).

Reorder the command-line arguments when building the hikari and hikari-unlocker
targets. Some compilers, like GCC, do not properly find symbols from libraries
unless the libraries are listed after the source files that reference them.
Libraries should always be listed last.

Add a ${DESTDIR} prefix to all install destinations, empty by default, to allow
installation into a chroot environment.

Drop the INSTALL_GROUP and associated -g flag to the install commands. Setting a
group explicitly is probably not useful, and can either cause breakage or
require manual specification for reasonable behavior. (In Void packaging
environments, for example, packages are built as an unprivileged user and the
default behavior of grabbing the user's primary group is not correct.) Files are
executable or readable by all users anyway and writable only by the owne