make workspace cycling attempt to focus last `focus_view`
This patch tries to focus the last `focus_view` during cycling. If that view is
no longer visible the first sheet view gets focused. This patch also makes
`hikari` focus the `focus_view` of a workspace when entering it with the cursor.

`focus_view` is no longer set to `NULL` when leaving a workspace.