fold migrate back into reset
Migrating a view has to happen instantly. Once a view leaves an output it needs
to be remove from all lists regarding that output and inserted into all of the
lists on the output we are migrating to. This operation should not be buffered
because otherwise the view can be somewhere in limbo between some outputs in a
state where one should not issue an operation on it. Anyway, if we are moving
the view via mouse it will never lose focus and therefore we could potentially
issue any operation violating invariants like that the focus view should never
be hidden.
By making the transition instant these "limbo states" cannot be achieved.