decouple view management from mapping
It currently can happen that an unmapped view can be shown on the workspace. If
the user switches to a sheet with an unmapped view, all linked sheet views get
shown, problematically this also applies to unmmaped views. There is no
distinction between `managed` an `mapped` state here so if a view gets unmapped
but not destroyed right away, it still happens to be managed by the sheet.

This patch separates the notion of `managed` and `mapped`. `unmapped` views get
removed from all the relevant lists to prevent the situation we are describing
in the previous paragraph.

`map` and `unmap` should be separate operations.