split up view configuration
Since `map` and `unmap` are now handled by views these functions now also need
to handle group assignment. An unmmapped view needs to be removed from its group
which might result in the group's deallocation. In case it's getting mapped
again we need to reload the view configuration to re-allocate the group.

This is done by factoring out the view configuration routine from the
configuration module and centralize view configuration in the view module via
`hikari_view_configure`. Configuration values are applied in this function
rather than in the resolve function for the `view_autoconf`. This also allows
for group allocation and deallocation in the `map` and `unmap` functions for
views.