destroy tiling information on `unmap`
Unmapping a tiled view leads to the tile being in detached state. Detached tiles
are tiles that do not belong to a layout. In general we want to avoid such tiles
to exist for a longer period of time (e.g. during resetting). Also unmapping all
the views from a layout will lead to the destruction of the layout.

In general, having detached tiles presented to the user will lead to surprising
effects, like being able to reset the geometry on a view that has no apparent
overriding geometry. This should be avoided.

Currently the best approach to deal with this is to just completely destroy the
tile. This also leads to overwriting the original geometry of the view leading
to some sort of "memory loss" for views that behave like that. This makes them a
bad candidate for tiling anyway.