fix focus issue on cycling
Cycling does not reset pointer focus properly so it can leave the pointer focus
on a different surface than the once we actually cycled to.

We now completely clear the focus on `hikari_workspace_focus_view` no matter
what and use the cursor to determine the new focus after cycling.