Rework force logic
Evacuating forced views needs to take into account that forced visible views are
not linked to the workspace related lists. With hidden visible views it is the
other way around. This needs to be handled when evacuating views in the case of
an output getting destroyed. Hidden forced views need to be removed from lists
and added correctly and vice versa.