Refactor operation logic
There is an issue with sending a resize event to surfaces which would not
require them to actually resize. In that case the serial of these operations
would be `0`. This way the commit logic would not recognize a surface update and
handle it wrongly.

This can be reproduced when trying to `fullscreen` a surface that already has
its maximum size.